[backend] Move bench-related entities to new access
