+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/benchmarkaccess/builder/NewMeasurement.java
@@ -1,9 +1,9 @@
-package de.aaaaaaah.velcom.backend.access.entities.benchmark;
+package de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.builder;
 
-import de.aaaaaaah.velcom.backend.access.entities.Measurement;
-import de.aaaaaaah.velcom.backend.access.entities.MeasurementError;
-import de.aaaaaaah.velcom.backend.access.entities.MeasurementValues;
-import de.aaaaaaah.velcom.backend.access.entities.RunId;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.Measurement;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.MeasurementError;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.MeasurementValues;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.RunId;
 import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Interpretation;
 import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Unit;
