+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/benchmarkaccess/entities/Measurement.java
@@ -1,4 +1,4 @@
-package de.aaaaaaah.velcom.backend.access.entities;
+package de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities;
 
 import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import de.aaaaaaah.velcom.shared.util.Either;
