+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/benchmarkaccess/entities/MeasurementError.java
@@ -1,4 +1,4 @@
-package de.aaaaaaah.velcom.backend.access.entities;
+package de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities;
 
 import java.util.Objects;
