+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/endpoints/GraphDetailEndpoint.java
@@ -3,10 +3,10 @@ package de.aaaaaaah.velcom.backend.restapi.endpoints;
 import static java.util.stream.Collectors.toCollection;
 import static java.util.stream.Collectors.toMap;
 
-import de.aaaaaaah.velcom.backend.access.entities.Measurement;
-import de.aaaaaaah.velcom.backend.access.entities.MeasurementValues;
-import de.aaaaaaah.velcom.backend.access.entities.Run;
 import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.BenchmarkReadAccess;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.Measurement;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.MeasurementValues;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.Run;
 import de.aaaaaaah.velcom.backend.newaccess.caches.LatestRunCache;
 import de.aaaaaaah.velcom.backend.newaccess.caches.RunCache;
 import de.aaaaaaah.velcom.backend.newaccess.committaccess.CommitReadAccess;
