+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/endpoints/utils/EndpointUtils.java
@@ -1,9 +1,9 @@
 package de.aaaaaaah.velcom.backend.restapi.endpoints.utils;
 
-import de.aaaaaaah.velcom.backend.access.entities.Run;
-import de.aaaaaaah.velcom.backend.access.entities.RunId;
 import de.aaaaaaah.velcom.backend.data.runcomparison.SignificanceFactors;
 import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.BenchmarkReadAccess;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.Run;
+import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.entities.RunId;
 import de.aaaaaaah.velcom.backend.newaccess.benchmarkaccess.exceptions.NoSuchRunException;
 import de.aaaaaaah.velcom.backend.newaccess.caches.LatestRunCache;
 import de.aaaaaaah.velcom.backend.newaccess.caches.RunCache;
